[WRFE-15](refactor): 폴더구조 변경 및 사용하지 않는 파일 삭제
