--- apps/front/wraffle-webview/src/entities/create/ui/index.ts
@@ -1,1 +1,0 @@
-export {ThumbnailCard} from './ThumbnailCard';
