+++ apps/front/wraffle-webview/src/entities/product-form/ui/ThumbnailCard.tsx
