--- apps/front/wraffle-webview/src/widgets/product-list/ui/index.ts
@@ -1,1 +1,0 @@
-export {TitleStep, DateStep, ImageStep, NoteStep} from './CreateList';
